+++ app/components/List/DeleteListButton.tsx
@@ -1,7 +1,6 @@
 import * as DropdownMenu from "@radix-ui/react-dropdown-menu";
 import styles from "../../styles/dropdownMenuContent.module.css";
-import { DotsCircleHorizontal } from "tabler-icons-react";
-import { Trash, CircleCheck } from "tabler-icons-react";
+import { DotsCircleHorizontal, Trash, CircleCheck } from "tabler-icons-react";
 import { useWatch } from "react-hook-form";
 
 type Props = {
@@ -13,7 +12,7 @@ type Item = {
   checked: boolean;
 };
 
-export default function DeleteMyMemoButton({ remove }: Props) {
+export default function DeleteListButton({ remove }: Props) {
   const fieldArray = useWatch({
     name: "list",
   });
